requirements.txt

# Core Web Frameworks
fastapi==0.111.0
uvicorn[standard]==0.30.1
streamlit==1.36.0

# HTTP and API
requests==2.32.3
httpx==0.27.0
python-dotenv==1.0.1

# Data Processing
pandas==2.2.2
numpy==2.0.1

# Machine Learning
scikit-learn==1.5.1
scipy==1.13.1

# Optional but recommended for production
# python-multipart==0.0.9  # For file uploads in FastAPI
# redis==5.0.7  # For distributed caching
# pytest==8.2.2  # For testing
# black==24.4.2  # For code formatting
# flake8==7.1.0  # For linting